# Migrate rlp from ElasticArray to SmallVec