--- a/util/src/lib.rs
+++ b/util/src/lib.rs
@@ -112,6 +112,7 @@ extern crate parking_lot;
 pub extern crate using_queue;
 pub extern crate table;
 extern crate ansi_term;
+extern crate smallvec;
 
 pub mod bloom;
 pub mod standard;