--- a/util/src/trie/triedbmut.rs
+++ b/util/src/trie/triedbmut.rs
@@ -24,7 +24,7 @@ use ::bytes::ToPretty;
 use ::nibbleslice::NibbleSlice;
 use ::rlp::{Rlp, RlpStream, View, Stream};
 
-use elastic_array::ElasticArray1024;
+use smallvec::SmallVec;
 
 use std::collections::{HashSet, VecDeque};
 use std::mem;
@@ -122,7 +122,7 @@ impl Node {
 
 	// encode a node to RLP
 	// TODO: parallelize
-	fn into_rlp<F>(self, mut child_cb: F) -> ElasticArray1024<u8>
+	fn into_rlp<F>(self, mut child_cb: F) -> SmallVec<[u8; 1024]>
 		where F: FnMut(NodeHandle, &mut RlpStream)
 	{
 		match self {