@rphmeier 2016-08-18
Thanks for doing this, @nipunn1313. Got a couple minor grumbles but mostly LGTM.