@nipunn1313 2016-08-19
Coolio. I fixed up some to_vec() and to to_owned() calls to be efficient/cleaner