@nipunn1313 2016-08-19
Trying to get smallvec to support Extend here. Looks like the extend inside smallvec is currently implemented as a push loop, but I'm sure that could also be optimized in the future.

https://github.com/servo/rust-smallvec/pull/27