@gavofyork 2016-08-19
introducing this kind of complexity in our codebase is a retrograde.

```
-	self.bytes.insert_slice(pos, &res);
+	for (idx, &item) in res.iter().enumerate() {
+		self.bytes.insert(pos + idx, item);
+	}
```

if we are to go with `SmallVec` over `ElasticArray`, then appropriate traits should be implemented to avoid such low-level algorithms straying into high-level code.