@nipunn1313 2016-08-19
Hey @gavofyork. I'm working on getting extend() added to the VecLike trait, but insert_slice may be a bit trickier. Let me see what I can do.