@nipunn1313 2016-08-20
https://github.com/servo/rust-smallvec/pull/28 and https://github.com/servo/rust-smallvec/pull/27 together should upgrade SmallVec to include the mechanisms we need (extend and insert_slice)

SmallVec just landed a change to implement Hash, so that handles the issue in kvdb, so this task seems quite tractable now.