@gavofyork 2016-08-23
code looks good, but fails tests.