@nipunn1313 2016-08-23
I do have a couple gripes that concern me before landing
1) SmallVec does not support insert_many/insert_slice just yet
2) SmallVec supports extend, but it doesn't use a memcpy like elasticarray. This is because SmallVec supports extending an iterable. SmallVec could support an optimized implementation for Vec<Copy> or Slice<Copy> if it used https://github.com/rust-lang/rfcs/blob/master/text/1210-impl-specialization.md, but specialization is an unstable feature right now.

I could probably get SmallVec to support InsertMany<Iterable>, but I think it'll be a little tougher to get support for the fast version.

That being said, I don't suspect that these code paths will get ruined by using the iterable implementation since it's typically copying <= 128 bytes at a time.

Want to point out these disadvantqges so we are aware of them.