@rphmeier 2016-08-23
Master:
```
running 7 tests
test bench_decode_nested_empty_lists ... bench:         537 ns/iter (+/- 12)
test bench_decode_u256_value         ... bench:         111 ns/iter (+/- 3)
test bench_decode_u64_value          ... bench:          50 ns/iter (+/- 2)
test bench_stream_1000_empty_lists   ... bench:      11,258 ns/iter (+/- 803)
test bench_stream_nested_empty_lists ... bench:         261 ns/iter (+/- 6)
test bench_stream_u256_value         ... bench:         565 ns/iter (+/- 12)
test bench_stream_u64_value          ... bench:         138 ns/iter (+/- 6)

test result: ok. 0 passed; 0 failed; 0 ignored; 7 measured

```

This PR:
```
running 7 tests
test bench_decode_nested_empty_lists ... bench:         533 ns/iter (+/- 8)
test bench_decode_u256_value         ... bench:         111 ns/iter (+/- 4)
test bench_decode_u64_value          ... bench:          49 ns/iter (+/- 1)
test bench_stream_1000_empty_lists   ... bench:       7,554 ns/iter (+/- 980)
test bench_stream_nested_empty_lists ... bench:         303 ns/iter (+/- 7)
test bench_stream_u256_value         ... bench:         655 ns/iter (+/- 20)
test bench_stream_u64_value          ... bench:         191 ns/iter (+/- 5)

test result: ok. 0 passed; 0 failed; 0 ignored; 7 measured
```

The empty lists case got a lot faster, but everything else got a bit slower. I think the `insert_many` optimization is probably the culprit. I'll throw together a crate for it.