@rphmeier 2016-08-23
Made an `InsertMany` trait but the benchmarks remain the same. Probably not a problem in practice but it remains to be seen: https://github.com/rphmeier/insert_many