@gavofyork 2016-08-24
@arkpar what do you think? `bench_stream_u64_value` got ~50% slower...