@nipunn1313 2016-08-24
@rphmeier. I believe that elastic array uses a memcpy for extend, but currently SmallVec uses a (somewhat inefficient) series of ptr::write in a loop. I don't think your insert_many crate is using that optimization and it definitely doesn't address all the calls to Extend.

I set up a PR to use memcpy for extend and insert_many in SmallVec.
https://github.com/servo/rust-smallvec/pull/29