@rphmeier 2016-08-24
The `trie` benches might be more indicative of practical effects (as they use RLP extensively). But they show a definite slowdown (even with `insert_many`)

master
```
$ rustup run nightly cargo bench -p ethcore-util --bench trie
running 13 tests
test sha3x10000                     ... bench:   6,555,022 ns/iter (+/- 110,064)
test trie_insertions_32_mir_1k      ... bench:   4,504,945 ns/iter (+/- 346,085)
test trie_insertions_32_ran_1k      ... bench:   4,503,505 ns/iter (+/- 250,628)
test trie_insertions_random_mid     ... bench:   3,910,607 ns/iter (+/- 228,881)
test trie_insertions_six_high       ... bench:   3,538,344 ns/iter (+/- 133,353)
test trie_insertions_six_low        ... bench:   8,409,129 ns/iter (+/- 416,024)
test trie_insertions_six_mid        ... bench:   5,216,624 ns/iter (+/- 307,361)
test triehash_insertions_32_mir_1k  ... bench:   1,939,029 ns/iter (+/- 26,818)
test triehash_insertions_32_ran_1k  ... bench:   1,943,441 ns/iter (+/- 322,224)
test triehash_insertions_random_mid ... bench:   1,208,044 ns/iter (+/- 31,692)
test triehash_insertions_six_high   ... bench:   1,386,069 ns/iter (+/- 24,725)
test triehash_insertions_six_low    ... bench:   1,969,671 ns/iter (+/- 37,346)
test triehash_insertions_six_mid    ... bench:   1,554,332 ns/iter (+/- 72,234)
```

this pr:
```
$ rustup run nightly cargo bench -p ethcore-util --bench trie
running 13 tests
test sha3x10000                     ... bench:   6,555,235 ns/iter (+/- 159,885)
test trie_insertions_32_mir_1k      ... bench:   4,951,804 ns/iter (+/- 566,059)
test trie_insertions_32_ran_1k      ... bench:   4,887,106 ns/iter (+/- 318,864)
test trie_insertions_random_mid     ... bench:   4,152,541 ns/iter (+/- 169,546)
test trie_insertions_six_high       ... bench:   3,779,139 ns/iter (+/- 192,456)
test trie_insertions_six_low        ... bench:   9,127,628 ns/iter (+/- 495,289)
test trie_insertions_six_mid        ... bench:   5,550,867 ns/iter (+/- 200,107)
test triehash_insertions_32_mir_1k  ... bench:   2,300,948 ns/iter (+/- 28,523)
test triehash_insertions_32_ran_1k  ... bench:   2,258,644 ns/iter (+/- 31,240)
test triehash_insertions_random_mid ... bench:   1,370,415 ns/iter (+/- 29,024)
test triehash_insertions_six_high   ... bench:   1,588,540 ns/iter (+/- 73,133)
test triehash_insertions_six_low    ... bench:   2,286,975 ns/iter (+/- 48,567)
test triehash_insertions_six_mid    ... bench:   1,788,652 ns/iter (+/- 28,400)
```

this pr with `insert_many`
```
$ rustup run nightly cargo bench -p ethcore-util --bench trie
running 13 tests
test sha3x10000                     ... bench:   6,541,748 ns/iter (+/- 248,270)
test trie_insertions_32_mir_1k      ... bench:   4,919,459 ns/iter (+/- 300,605)
test trie_insertions_32_ran_1k      ... bench:   4,864,872 ns/iter (+/- 251,946)
test trie_insertions_random_mid     ... bench:   4,087,234 ns/iter (+/- 160,874)
test trie_insertions_six_high       ... bench:   3,757,468 ns/iter (+/- 160,793)
test trie_insertions_six_low        ... bench:   8,852,989 ns/iter (+/- 455,819)
test trie_insertions_six_mid        ... bench:   5,462,281 ns/iter (+/- 156,619)
test triehash_insertions_32_mir_1k  ... bench:   2,286,986 ns/iter (+/- 21,361)
test triehash_insertions_32_ran_1k  ... bench:   2,249,132 ns/iter (+/- 74,144)
test triehash_insertions_random_mid ... bench:   1,376,235 ns/iter (+/- 23,817)
test triehash_insertions_six_high   ... bench:   1,592,853 ns/iter (+/- 63,619)
test triehash_insertions_six_low    ... bench:   2,318,934 ns/iter (+/- 44,270)
test triehash_insertions_six_mid    ... bench:   1,796,758 ns/iter (+/- 30,831)
```